+++ nodeconductor/cost_tracking/handlers.py
@@ -4,8 +4,9 @@ import logging
 from dateutil.relativedelta import relativedelta
 
 from nodeconductor.core.tasks import send_task
+from nodeconductor.core.utils import serialize_instance
 from nodeconductor.cost_tracking import exceptions, models, CostTrackingRegister
-from nodeconductor.structure.models import ResourceMixin, Resource
+from nodeconductor.cost_tracking.models import PayableMixin
 from nodeconductor.structure import SupportedServices, ServiceBackendNotImplemented, ServiceBackendError
 
 logger = logging.getLogger(__name__)
@@ -60,7 +61,7 @@ def copy_threshold_from_previous_price_estimate(sender, instance, created=False,
 
 def update_projected_estimate(sender, instance, **kwargs):
     send_task('cost_tracking', 'update_projected_estimate')(
-        resource_str=instance.to_string())
+        serialized_resource=serialize_instance(instance))
 
 
 def update_price_estimate_ancestors(sender, instance, created=False, **kwargs):
@@ -126,7 +127,7 @@ def delete_price_estimate_on_scope_deletion(sender, instance, **kwargs):
     # if scope is Resource:
     #    delete -- add metadata about deleted resource, set object_id to NULL
     #    unlink -- delete all related estimates
-    if isinstance(instance, tuple(ResourceMixin.get_all_models())):
+    if isinstance(instance, tuple(PayableMixin.get_all_models())):
         if getattr(instance, 'PERFORM_UNLINK', False):
             models.PriceEstimate.delete_estimates_for_resource(instance)
         else: