+++ CHANGELOG.rst
@@ -3,7 +3,12 @@ CHANGELOG
 
 Release NEXT
 ------------
-Nothing yet.
+- Nothing yet.
+
+Release 0.102.3
+---------------
+- Rename PaidResource to PayableMixin. Track PriceEstimate for PayableMixin.
+- Allow executing custom actions via templates after provision.
 
 Release 0.102.2
 ---------------